Combine bool and non bool cases files
