+++ src/Octopus.OpenFeature.Provider.SpecificationTests/Cases.cs
@@ -3,19 +3,24 @@ using System.Text.Json;
 
 namespace Octopus.OpenFeature.Provider.SpecificationTests;
 
-public class Cases : IEnumerable<object[]>
+public interface IFixture<T>
+{
+    JsonElement Response { get; }
+    T[] Cases { get; }
+}
+
+public abstract class Cases<TFixture, TCase>(string directory) : IEnumerable<object[]> where TFixture : IFixture<TCase>
 {
     public IEnumerator<object[]> GetEnumerator()
     {
-        var jsonFiles = Directory.EnumerateFiles("Fixtures/boolean", "*.json");
+        var jsonFiles = Directory.EnumerateFiles(directory, "*.json");
 
         foreach (var jsonFile in jsonFiles)
         {
             var json = File.ReadAllText(jsonFile);
-            var data = JsonSerializer.Deserialize<Fixture>(json, JsonSerializerOptions.Web)!;
+            var data = JsonSerializer.Deserialize<TFixture>(json, JsonSerializerOptions.Web)!;
 
             var responseJson = data.Response.GetRawText();
-            var file = Path.GetFileName(jsonFile);
 
             foreach (var c in data.Cases)
             {
@@ -30,10 +35,12 @@ public class Cases : IEnumerable<object[]>
     }
 }
 
+public class BooleanCases() : Cases<Fixture, FixtureCase>("Fixtures/boolean");
+
 public record Fixture(
     JsonElement Response,
     FixtureCase[] Cases
-);
+) : IFixture<FixtureCase>;
 
 public record FixtureCase(
     string Description,
@@ -41,8 +48,7 @@ public record FixtureCase(
     FixtureExpected Expected
 );
 
-public record FixtureConfiguration(
-    string Slug,
+public record FixtureConfiguration(string Slug,
     bool DefaultValue,
     Dictionary<string, string>? Context
 );
@@ -50,4 +56,27 @@ public record FixtureConfiguration(
 public record FixtureExpected(
     bool Value,
     string? ErrorCode = null
+);
+
+public class NonBooleanCases() : Cases<NonBooleanFixture, NonBooleanFixtureCase>("Fixtures/non-boolean");
+
+public record NonBooleanFixture(
+    JsonElement Response,
+    NonBooleanFixtureCase[] Cases
+) : IFixture<NonBooleanFixtureCase>;
+
+public record NonBooleanFixtureCase(
+    string Description,
+    NonBooleanFixtureConfiguration Configuration,
+    NonBooleanFixtureExpected Expected
+);
+
+public record NonBooleanFixtureConfiguration(
+    string Slug,
+    string FlagType
+);
+
+public record NonBooleanFixtureExpected(
+    JsonElement Value,
+    string ErrorCode
 );
